third_party/blink/web_tests/platform/generic/fast/filesystem/null-arguments-expected.txt

This test tries calling various filesystem functions with null arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Successfully obtained FileSystem: file__0:Temporary
PASS fileSystem.root.moveTo(null) threw exception TypeError: Failed to execute 'moveTo' on 'Entry': parameter 1 is not of type 'DirectoryEntry'..
PASS fileSystem.root.copyTo(null) threw exception TypeError: Failed to execute 'copyTo' on 'Entry': parameter 1 is not of type 'DirectoryEntry'..
PASS writer.error is null
PASS writer.write(null) threw exception TypeError: Failed to execute 'write' on 'FileWriter': parameter 1 is not of type 'Blob'..
PASS writer.error is null
PASS successfullyParsed is true

TEST COMPLETE
